backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.2
psycopg2-binary==2.9.9

# Auth
supabase==2.5.0
python-jose[cryptography]==3.3.0

# Task queue
celery==5.4.0
redis==5.0.4

# AI
# anthropic==0.28.0
google-generativeai==0.7.2

# File parsing
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
Pillow==10.3.0

# Utils
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0